backend/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
requests==2.32.3
neo4j==5.23.0
fastapi==0.115.0
uvicorn==0.30.0
structlog==24.1.0
openai==1.51.2
instructor==1.3.4
networkx==3.4.2
prometheus-client==0.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
arq==0.26
redis==4.5.5
httpx==0.27.2
qdrant-client==1.6.6
strawberry-graphql==0.211.0
pytest==8.3.2
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
